requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy==1.24.3
pandas==2.0.2
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3